added modules and updates some fixes

+++ inc/plugins/hello.php
@@ -13,7 +13,7 @@ if(!defined('IN_MYBB'))
 {
 	die('This file cannot be accessed directly.');
 }
-
+	
 // cache templates - this is important when it comes to performance
 // THIS_SCRIPT is defined by some of the MyBB scripts, including index.php
 if(defined('THIS_SCRIPT'))
@@ -74,9 +74,9 @@ function hello_info()
 	return array(
 		'name'			=> 'Hello World!',
 		'description'	=> $lang->hello_desc,
-		'website'		=> 'https://mybb.com',
+		'website'		=> 'http://mybb.com',
 		'author'		=> 'MyBB Group',
-		'authorsite'	=> 'https://mybb.com',
+		'authorsite'	=> 'http://www.mybb.com',
 		'version'		=> '2.0',
 		'compatibility'	=> '18*',
 		'codename'		=> 'hello'
@@ -318,13 +318,13 @@ function hello_activate()
 
 	// Delete deprecated entries.
 	$db->delete_query('settings', "gid='{$gid}' AND description='HELLODELETEMARKER'");
-
+	
 	// This is required so it updates the settings.php file as well and not only the database - they must be synchronized!
 	rebuild_settings();
 
 	// Include this file because it is where find_replace_templatesets is defined
 	require_once MYBB_ROOT.'inc/adminfunctions_templates.php';
-
+	
 	// Edit the index template and add our variable to above {$forums}
 	find_replace_templatesets('index', '#'.preg_quote('{$forums}').'#', "{\$hello}\n{\$forums}");
 }
@@ -339,7 +339,7 @@ function hello_activate()
 function hello_deactivate()
 {
 	require_once MYBB_ROOT.'inc/adminfunctions_templates.php';
-
+	
 	// remove template edits
 	find_replace_templatesets('index', '#'.preg_quote('{$hello}').'#', '');
 }
@@ -354,35 +354,17 @@ function hello_install()
 {
 	global $db;
 
-	// Create our table collation
+	// Create our entries table
 	$collation = $db->build_create_table_collation();
 
-	// Create table if it doesn't exist already
+	// create table if it doesn't exist already
 	if(!$db->table_exists('hello_messages'))
 	{
-		switch($db->type)
-		{
-			case "pgsql":
-				$db->write_query("CREATE TABLE ".TABLE_PREFIX."hello_messages (
-					mid serial,
-					message varchar(100) NOT NULL default '',
-					PRIMARY KEY (mid)
-				);");
-				break;
-			case "sqlite":
-				$db->write_query("CREATE TABLE ".TABLE_PREFIX."hello_messages (
-					mid INTEGER PRIMARY KEY,
-					message varchar(100) NOT NULL default ''
-				);");
-				break;
-			default:
-				$db->write_query("CREATE TABLE ".TABLE_PREFIX."hello_messages (
-					mid int unsigned NOT NULL auto_increment,
-					message varchar(100) NOT NULL default '',
-					PRIMARY KEY (mid)
-				) ENGINE=MyISAM{$collation};");
-				break;
-		}
+		$db->write_query("CREATE TABLE `".TABLE_PREFIX."hello_messages` (
+			`mid` int(10) UNSIGNED NOT NULL auto_increment,
+			`message` varchar(100) NOT NULL default '',
+			PRIMARY KEY  (`mid`)
+		) ENGINE=MyISAM{$collation}");
 	}
 }
 
@@ -417,22 +399,38 @@ function hello_uninstall()
 		$page->output_confirm_action('index.php?module=config-plugins&action=deactivate&uninstall=1&plugin=hello', $lang->hello_uninstall_message, $lang->hello_uninstall);
 	}
 
-	// Delete template groups.
-	$db->delete_query('templategroups', "prefix='hello'");
+	// remove our templates group
+	// Query the template groups
+	$query = $db->simple_select('templategroups', 'prefix', "prefix='hello'");
 
-	// Delete templates belonging to template groups.
-	$db->delete_query('templates', "title='hello' OR title LIKE 'hello_%'");
+	// Build where string for templates
+	$sqlwhere = array();
 
-	// Delete settings group
+	while($prefix = $db->fetch_field($query, 'prefix'))
+	{
+		$tprefix = $db->escape_string($prefix);
+		$sqlwhere[] = "title='{$tprefix}' OR title LIKE '{$tprefix}=_%' ESCAPE '='";
+	}
+
+	if($sqlwhere) // else there are no groups to delete
+	{
+		// Delete template groups.
+		$db->delete_query('templategroups', "prefix='hello'");
+
+		// Delete templates belonging to template groups.
+		$db->delete_query('templates', implode(' OR ', $sqlwhere));
+	}
+
+	// delete settings group
 	$db->delete_query('settinggroups', "name='hello'");
 
-	// Remove the settings
+	// remove settings
 	$db->delete_query('settings', "name IN ('hello_display1','hello_display2')");
 
 	// This is required so it updates the settings.php file as well and not only the database - they must be synchronized!
 	rebuild_settings();
 
-	// Drop tables if desired
+	// drop tables if desired
 	if(!isset($mybb->input['no']))
 	{
 		$db->drop_table('hello_messages');
